Share a common fake logreader across tests
